draft edge boostrap logic
